enum refactor, whew
