+++ crates/lib/database_queries/src/queries/vocoder/get_vocoder_model.rs
@@ -8,8 +8,8 @@ use chrono::{DateTime, Utc};
 use container_common::anyhow_result::AnyhowResult;
 use crate::helpers::boolean_converters::i8_to_bool;
 use log::warn;
-use reusable_types::entity_visibility::EntityVisibility;
-use reusable_types::vocoder_type::VocoderType;
+use reusable_types::db::enums::entity_visibility::EntityVisibility;
+use reusable_types::db::enums::vocoder_type::VocoderType;
 use sqlx::MySqlPool;
 
 // FIXME: This is the old style of query scoping and shouldn't be copied.
@@ -120,7 +120,7 @@ async fn select_including_deleted(
         r#"
 SELECT
     vocoder.token as vocoder_token,
-    vocoder.vocoder_type as `vocoder_type: reusable_types::vocoder_type::VocoderType`,
+    vocoder.vocoder_type as `vocoder_type: reusable_types::db::enums::vocoder_type::VocoderType`,
 
     vocoder.creator_user_token,
     users.username as creator_username,
@@ -137,7 +137,7 @@ SELECT
     vocoder.is_mod_disabled_from_author_use,
     vocoder.is_mod_author_editing_locked,
 
-    vocoder.creator_set_visibility as `creator_set_visibility: reusable_types::entity_visibility::EntityVisibility`,
+    vocoder.creator_set_visibility as `creator_set_visibility: reusable_types::db::enums::entity_visibility::EntityVisibility`,
 
     vocoder.created_at,
     vocoder.updated_at,
@@ -168,7 +168,7 @@ async fn select_without_deleted(
         r#"
 SELECT
     vocoder.token as vocoder_token,
-    vocoder.vocoder_type as `vocoder_type: reusable_types::vocoder_type::VocoderType`,
+    vocoder.vocoder_type as `vocoder_type: reusable_types::db::enums::vocoder_type::VocoderType`,
 
     vocoder.creator_user_token,
     users.username as creator_username,
@@ -185,7 +185,7 @@ SELECT
     vocoder.is_mod_disabled_from_author_use,
     vocoder.is_mod_author_editing_locked,
 
-    vocoder.creator_set_visibility as `creator_set_visibility: reusable_types::entity_visibility::EntityVisibility`,
+    vocoder.creator_set_visibility as `creator_set_visibility: reusable_types::db::enums::entity_visibility::EntityVisibility`,
 
     vocoder.created_at,
     vocoder.updated_at,
