+++ crates/lib/database_queries/src/queries/vocoder/list_vocoder_models.rs
@@ -8,7 +8,7 @@ use chrono::{DateTime, Utc};
 use container_common::anyhow_result::AnyhowResult;
 use crate::helpers::boolean_converters::i8_to_bool;
 use log::{warn, info};
-use reusable_types::vocoder_type::VocoderType;
+use reusable_types::db::enums::vocoder_type::VocoderType;
 use sqlx::MySqlPool;
 
 // FIXME: This is the old style of query scoping and shouldn't be copied.
@@ -115,7 +115,7 @@ async fn list_vocoder_models_for_all_creators(
         r#"
 SELECT
     vocoder.token as vocoder_token,
-    vocoder.vocoder_type as `vocoder_type: reusable_types::vocoder_type::VocoderType`,
+    vocoder.vocoder_type as `vocoder_type: reusable_types::db::enums::vocoder_type::VocoderType`,
     vocoder.creator_user_token,
     users.username as creator_username,
     users.display_name as creator_display_name,
@@ -147,7 +147,7 @@ WHERE
         r#"
 SELECT
     vocoder.token as vocoder_token,
-    vocoder.vocoder_type as `vocoder_type: reusable_types::vocoder_type::VocoderType`,
+    vocoder.vocoder_type as `vocoder_type: reusable_types::db::enums::vocoder_type::VocoderType`,
     vocoder.creator_user_token,
     users.username as creator_username,
     users.display_name as creator_display_name,
@@ -187,7 +187,7 @@ async fn list_vocoder_models_creator_scoped(
         r#"
 SELECT
     vocoder.token as vocoder_token,
-    vocoder.vocoder_type as `vocoder_type: reusable_types::vocoder_type::VocoderType`,
+    vocoder.vocoder_type as `vocoder_type: reusable_types::db::enums::vocoder_type::VocoderType`,
     vocoder.creator_user_token,
     users.username as creator_username,
     users.display_name as creator_display_name,
@@ -221,7 +221,7 @@ WHERE
         r#"
 SELECT
     vocoder.token as vocoder_token,
-    vocoder.vocoder_type as `vocoder_type: reusable_types::vocoder_type::VocoderType`,
+    vocoder.vocoder_type as `vocoder_type: reusable_types::db::enums::vocoder_type::VocoderType`,
     vocoder.creator_user_token,
     users.username as creator_username,
     users.display_name as creator_display_name,
