+++ crates/lib/reusable_types/src/db/enums/mod.rs
@@ -1,0 +1,7 @@
+//! This is meant to handle `enum` or `enum`-like columns in the database.
+//! These are not Rust-like enums (sum types) that can have internal payloads.
+
+pub mod entity_visibility;
+pub mod generic_download_type;
+pub mod generic_inference_type;
+pub mod vocoder_type;
