+++ crates/service/web/storyteller_web/src/http_server/endpoints/download_job/enqueue_generic_download.rs
@@ -16,8 +16,8 @@ use http_server_common::request::get_request_ip::get_request_ip;
 use http_server_common::response::serialize_as_json_error::serialize_as_json_error;
 use log::{info, warn, log};
 use regex::Regex;
-use reusable_types::entity_visibility::EntityVisibility;
-use reusable_types::generic_download_type::GenericDownloadType;
+use reusable_types::db::enums::entity_visibility::EntityVisibility;
+use reusable_types::db::enums::generic_download_type::GenericDownloadType;
 use sqlx::error::DatabaseError;
 use sqlx::error::Error::Database;
 use sqlx::mysql::MySqlDatabaseError;
